+++ service/src/main/java/com/epam/lab/service/BaseService.java
@@ -52,5 +52,5 @@ public interface BaseService<T> {
      *
      * @return the int
      */
-    int countAll();
+    long countAll();
 }
